refactor(layout): 측정·렌더 공유 줄바꿈 코어를 HwpLineBreaker로 분리한다 (#80)
측정(HwpParagraphLayout.layout)과 렌더(HwpDrawnTextLayout.lines)가 실제로
공유하는 코어의 이름이 "그려진 텍스트 레이아웃"이라, 측정 경로를 따라온
사람에게 이 코어가 왜 거기 있는지 드러나지 않았다. 공유 멤버 5개를
Text/HwpLineBreaker.swift로 옮긴다 — FrameChunk · nextFrameChunk ·
paragraphStyle(in:at:) · paragraphCGFloat · availableLineWidth.

- resumeBaseline·fallbackLineAdvance는 옮기지 않는다. 둘 다 호출자가
  lines()뿐인 렌더 전용 헬퍼이고, 측정은 makeLineFrames가 origin 델타로
  자체 처리한다. 공유 코어에 렌더 전용 멤버를 넣으면 이름이 다시 거짓말을
  한다.
- 접근 수준은 그대로다 (internal 4 + private 1). 남는 fallbackLineAdvance가
  paragraphStyle·paragraphCGFloat를 계속 부르므로 둘은 internal이어야 한다.
- 파일 밖 호출부는 HwpParagraphLayout.swift:184 하나뿐이었다.
- 순수 이동에 더한 것은 긴 시그니처 2개 줄바꿈뿐이다 (line_length 경고 해소).

문서화가 이 조각의 절반이다 — nextFrameChunk·FrameChunk는 저장소 전체
*.md에서 grep 0건이라 여러 라운드가 쌓은 계약이 지식 베이스에 없었다.
Sources/HwpKitCore/AGENTS.md에 새 절을 두어 계약 넷(줄 예산 절단 · 미완
마지막 줄 이월 · 한 시각 줄 재프레이밍 · keepCount 하드 상한)과 "렌더
전용 멤버 금지"를 적고, 양쪽 정렬 항목에서 상호 참조한다.

검증: swift test 전 스위트 통과 (렌더 가드 4층 재기록 없음).
swiftlint 저장소 경고 192 → 188 (HwpDrawnTextLayout.swift 502 → 382줄로
file_length·type_body_length 경고 해소, 이동한 2건은 신설 파일로 승계).

Refs #80
